services/product-service/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
beanie==1.23.6
motor==3.3.2
pymongo==4.6.0

# gRPC
grpcio==1.59.3
grpcio-tools==1.59.3
grpcio-health-checking==1.59.3
protobuf==4.25.1

# Async support
asyncio-mqtt==0.16.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Logging and monitoring
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
